+++ pom.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
-  Copyright © 2017 Cask Data, Inc.
+  Copyright © 2017-2019 Cask Data, Inc.
 
   Licensed under the Apache License, Version 2.0 (the "License"); you may not
   use this file except in compliance with the License. You may obtain a copy of
@@ -20,18 +20,18 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
 
-  <groupId>co.cask.plugins</groupId>
+  <groupId>io.cdap.plugin</groupId>
   <artifactId>condition-plugins</artifactId>
   <version>1.3.0-SNAPSHOT</version>
   <name>Condition Plugin</name>
   <description>A plugin for making control flow decision within a pipeline</description>
   <packaging>jar</packaging>
 
-  <url>https://github.com/hydrator/condition-plugins</url>
+  <url>https://github.com/data-integrations/condition-plugins</url>
 
   <organization>
-    <name>Cask Data, Inc.</name>
-    <url>http://cask.co</url>
+    <name>CDAP</name>
+    <url>http://cdap.io</url>
   </organization>
 
   <licenses>
@@ -43,17 +43,17 @@
 
   <developers>
     <developer>
-      <name>Cask Data</name>
-      <email>cask-dev@googlegroups.com</email>
-      <organization>Cask Data, Inc.</organization>
-      <organizationUrl>http://www.cdap.io</organizationUrl>
+      <name>CDAP</name>
+      <email>cdap-dev@googlegroups.com</email>
+      <organization>CDAP</organization>
+      <organizationUrl>http://cdap.io</organizationUrl>
     </developer>
   </developers>
 
   <scm>
-    <connection>scm:git:https://github.com/hydrator/condition-plugins.git</connection>
-    <developerConnection>scm:git:git@github.com:hydrator/condition-plugins.git</developerConnection>
-    <url>https://github.com/hydrator/condition-plugins.git</url>
+    <connection>scm:git:https://github.com/data-integrations/condition-plugins.git</connection>
+    <developerConnection>scm:git:git@github.com:data-integrations/condition-plugins.git</developerConnection>
+    <url>https://github.com/data-integrations/condition-plugins.git</url>
     <tag>HEAD</tag>
   </scm>
 
@@ -66,10 +66,6 @@
       <id>sonatype.snapshots</id>
       <url>https://oss.sonatype.org/content/repositories/snapshots</url>
     </snapshotRepository>
-    <site>
-      <id>cdap</id>
-      <url>http://cdap.io</url>
-    </site>
   </distributionManagement>
 
   <issueManagement>
@@ -121,13 +117,13 @@
       <version>${slf4j.version}</version>
     </dependency>
     <dependency>
-      <groupId>co.cask.cdap</groupId>
+      <groupId>io.cdap.cdap</groupId>
       <artifactId>cdap-api</artifactId>
       <version>${cdap.version}</version>
       <scope>provided</scope>
     </dependency>
     <dependency>
-      <groupId>co.cask.cdap</groupId>
+      <groupId>io.cdap.cdap</groupId>
       <artifactId>cdap-etl-api</artifactId>
       <version>${cdap.version}</version>
       <scope>provided</scope>
@@ -138,7 +134,7 @@
       <version>${guava.version}</version>
     </dependency>
     <dependency>
-      <groupId>co.cask.cdap</groupId>
+      <groupId>io.cdap.cdap</groupId>
       <artifactId>cdap-unit-test</artifactId>
       <version>${cdap.version}</version>
       <scope>test</scope>
@@ -257,7 +253,7 @@
             <Embed-Transitive>true</Embed-Transitive>
             <Embed-Directory>lib</Embed-Directory>
             <!--Only @Plugin classes in the export packages will be included as plugin-->
-            <_exportcontents>co.cask.plugins.*</_exportcontents>
+            <_exportcontents>io.cdap.plugin.condition.*</_exportcontents>
           </instructions>
         </configuration>
         <executions>
@@ -270,9 +266,9 @@
         </executions>
       </plugin>
       <plugin>
-        <groupId>co.cask</groupId>
+        <groupId>io.cdap</groupId>
         <artifactId>cdap-maven-plugin</artifactId>
-        <version>1.0.1</version>
+        <version>1.1.0</version>
         <configuration>
           <cdapArtifacts>
             <parent>system:cdap-data-pipeline[5.0.0-SNAPSHOT,9.0.0-SNAPSHOT)</parent>
@@ -329,7 +325,7 @@
                 </links>
                 <doctitle>${project.name} ${project.version}</doctitle>
                 <bottom>
-                  <![CDATA[Copyright &#169; {currentYear} <a href="http://cask.co" target="_blank">Cask Data, Inc.</a> Licensed under the <a href="http://www.apache.org/licenses/LICENSE-2.0" target="_blank">Apache License, Version 2.0</a>.]]>
+                  <![CDATA[Copyright &#169; {currentYear} <a href="http://cdap.io" target="_blank">CDAP</a> Licensed under the <a href="http://www.apache.org/licenses/LICENSE-2.0" target="_blank">Apache License, Version 2.0</a>.]]>
                 </bottom>
               </configuration>
               <executions>
@@ -351,7 +347,6 @@
               <configuration>
                 <nexusUrl>https://oss.sonatype.org</nexusUrl>
                 <serverId>sonatype.release</serverId>
-                <stagingProfileId>655dc88dc770c3</stagingProfileId>
               </configuration>
             </plugin>
