+++ src/main/java/io/cdap/plugin/condition/Global.java
@@ -14,7 +14,7 @@
  *  the License.
  */
 
-package co.cask.plugins;
+package io.cdap.plugin.condition;
 
 
 /**
